Create README.md
This is Meshack Wanyama's Porfolio website showcasing my background, skills, and interest as a future full-stack developer. Key sections in this website include educational background, About me
section, the programming languages that am familiar with and a contact form form for those who wish to collaborate with me or enquire anything. To create this website, I used mostly html and a
little bit of css. I hope to learn more about the latter since am just but a newbie.